% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citations.r
\name{cite_pkg}
\alias{cite_pkg}
\title{Cite a package in markdown}
\usage{
cite_pkg(pkg)
}
\arguments{
\item{pkg}{The name of the package to cide}
}
\value{
a markdown string
}
\description{
Cite a package in markdown
}
\examples{
cite_pkg("knitr")
}
