% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lncs.r
\name{generate_environment}
\alias{generate_environment}
\title{generate LNCS Environments with a counter}
\usage{
generate_environment(env = "Theorem")
}
\arguments{
\item{env}{The name of the environment (e.g., "Theorem", "Lemma", etc.)}
}
\value{
A function that formats output for the environment
}
\description{
generate LNCS Environments with a counter
}
\examples{
theorem <- generate_environment("Theorem")
}
