% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citations.r
\name{line_cite}
\alias{line_cite}
\title{Create a line that cites packges}
\usage{
line_cite(pkgs)
}
\arguments{
\item{pkgs}{a Vector containing package names to be cited}
}
\value{
a markdown string
}
\description{
Create a line that cites packges
}
\examples{
line_cite(c("rmdtemplates", "knitr"))
}
