% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citations.r
\name{list_cite}
\alias{list_cite}
\title{Create a list that cites packges}
\usage{
list_cite(pkgs)
}
\arguments{
\item{pkgs}{a Vector containing package names to be cited}
}
\value{
a markdown string
}
\description{
Create a list that cites packges
}
\examples{
list_cite(c("rmdtemplates", "knitr"))
}
