% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figurecaptions.R
\name{ref}
\alias{ref}
\title{A referencing function that wraps figure labels for both html and latex output}
\usage{
ref(label, text = "", output = knitr::opts_knit$get("rmarkdown.pandoc.to"))
}
\arguments{
\item{label}{the figure label}

\item{text}{An optional reference text}

\item{output}{the output format either "latex" or "html"}
}
\value{
a markdown string formatted matching the output format
}
\description{
A referencing function that wraps figure labels for both html and latex output
}
\examples{
ref("fig:fig1")
}
