% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citations.r
\name{write_bib}
\alias{write_bib}
\title{Write the bib-tex file from alist of package-strings}
\usage{
write_bib(pkgs)
}
\arguments{
\item{pkgs}{a Vector containing package names to be cited}
}
\value{
nothing
}
\description{
Write the bib-tex file from alist of package-strings
}
\examples{
write_bib(c("rmdtemplates", "knitr"))
}
